add exception handler, customized exception classes

+++ src/main/java/com/ecommerce/booksale/book/BookRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long>, PagingAndSortingRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer>, PagingAndSortingRepository<Book, Integer> {
     // find All Books
     List<Book> findAll();
